Fix minor issues for no action and new dicom header reading

+++ weasis-dicom/weasis-dicom-codec/src/main/java/org/weasis/dicom/codec/DicomImageElement.java
@@ -734,8 +734,8 @@ public class DicomImageElement extends ImageElement {
                     // If no sliceTickness: set 0, sliceTickness is only use in IntersectVolume
                     // Multiply rows and columns by getZoomScale() to have square pixel image size
                     return new GeometryOfSlice(new double[] { imgOr[0], imgOr[1], imgOr[2] }, new double[] { imgOr[3],
-                        imgOr[4], imgOr[5] }, pos, spacing, sliceTickness == null ? 0.0 : sliceTickness.doubleValue(),
-                        new double[] { rows * getRescaleY(), columns * getRescaleX(), 1 });
+                        imgOr[4], imgOr[5] }, pos, spacing, sliceTickness == null ? 0.0 : sliceTickness, new double[] {
+                        rows * getRescaleY(), columns * getRescaleX(), 1 });
                 }
             }
         }
